Add links to Articles, UX/UI and Podcast
Articles - JSON in Swift
Note: Changed the title that appears on blog post: SWIFT JSON SHOOT-OUT

UX/UI - iOS Video Backgrounds
Note: I have a few doubts where this link should be or if it should be at all. Although the article explains how to do it, it's about this new trend (?), so I thought it could be in interesting to include.
Whatever you decide is perfect.

Podcast - Bonus Episode 1: Apple Watch Watch 2015