Add tool
This is my tool 😄

Flawless App helps macOS developers to easily spot visual differences between design and implementation right inside iOS simulator. With Flawless mobile teams can make polished and nice looking apps very fast. Tool has overlay\split comparison modes and support of long\landscape design screens comparison

We made Flawless App together with Ahmed & our Ukrainian team. We spent 2 years to make, test and polish it.  Now it’s easy-to-use well-tested tool with full support. 

We want to make Flawless App accessible for developers with different financial situation, that's why we put a fair price for it ($15 for life-long use). Beta-version was free and it will be free for early users forever. I know, that community prefer OS or free products, as well as we do. But, like many tools, we have to go to the commercial product. For a very long time we work on Flawless full time and we want to continue improving it without investments (as long, as we can at least).

I would really appreciate if you add it :)