Add business article
I made a list of 17 great ladies & iOS engineers, whos Twitter I love! It will be a useful follow list for all newbie iOS devs and for those who wish to support diversity in tech. Not sure, if it's a better fit for business or general section. Please correct me, if I'm wrong.