Add article - "Outsourcing Requests" (#1377)
* Add article - "Outsourcing Requests"

A few days ago I wrote an article on a practice I began doing in my apps, where I would move all external API requests (even the simple ones) outside of the app and up to the backend. I explain my reasoning for doing so and use a real-life case to show the practice in action.

(official excerpt from my website): "As applications grow more and more complex, they may have to call an increasing number of external web APIs. As we build more features, we may depend on services that we do not control, for the sake of gaining functionality we wouldn’t be able to implement ourselves. That’s great! What’s not so great is when one of those services goes down, or has a major breaking change, or goes away altogether. Then, if we have not planned correctly, we may completely break our app with no immediate fix available. This is why I propose that to retain complete control over user experience, it is important that we maintain complete control over every external web call that goes in and out of our app."

* Update Week269.md